Create TripsController for working with Trips
Hoo boy... Lots of new syntax to learn.

Looks like this controller is automatically used when visiting the API
URL `/api/TripsController` or something like that... Also, it looks like
there are methods that will be called automatically depending on the
type of request, and what data is included.

For instance, when there's a GET request, the `Get()` method that
returns an `IEnumerable<Trip>` is called automatically. However, if
there's a GET request that includes an `id`, the `Get()` method that
returns a `Trip` is called instead.

Likewise there are methods automatically called for POST, PUT, and
DELETE requests.

Pick up on video at 42:30 to continue.